# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "World UZ",
        "description": "Documentation for World Uz Company",
        "contact": {
            "email": "rustamovramziddin7@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/country": {
            "get": {
                "tags": [
                    "Country"
                ],
                "summary": "Return only name and image for the  first page to avoid large date",
                "description": "Return only name and image for the  first page",
                "operationId": "fc919d08df191af41faa4e06106a61c3",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Country"
                ],
                "summary": "Post a new data",
                "description": "Post new coutry  data",
                "operationId": "9268288b99685cf1f05fdcfa283939db",
                "requestBody": {
                    "description": "Pass Country   credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "AQSH"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": "image"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/v1/country/{country}": {
            "get": {
                "tags": [
                    "Country"
                ],
                "summary": "Get one ",
                "description": "Return all date related to ID{country id}",
                "operationId": "981c7edc221587967b9b433fbdc522d6",
                "parameters": [
                    {
                        "name": "country",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Country"
                ],
                "summary": "Get one and Delete related to country id",
                "description": "Return  date related to ID of the country",
                "operationId": "15454d9e8580ce23a03ce453ff024ad5",
                "parameters": [
                    {
                        "name": "country",
                        "in": "path",
                        "description": "put country id and try to delete ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/v1/countries/{country}/university": {
            "get": {
                "tags": [
                    "Country"
                ],
                "summary": "Get one with relation table universities , put country id and get universities ",
                "description": "Return all date related to ID{country id}",
                "operationId": "3ae321ed959e339072ad6247e51ef400",
                "parameters": [
                    {
                        "name": "country",
                        "in": "path",
                        "description": "ID of the country",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}